requirements.txt
zope.component [zcml]

# rdflib
# mercurial
#python-hglib
# lxml
# icc.ngs
